requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
httpx==0.26.0
pyjwt==2.8.1
cryptography==42.0.4
